refactor: move run package commands and helper files to cli package
