+++ packages/cli/src/commands/console.ts
@@ -1,8 +1,8 @@
 // tslint:disable:file-name-casing
 import {Command, flags} from '@heroku-cli/command'
 import {DynoSizeCompletion} from '@heroku-cli/command/lib/completions'
-import Dyno from '../lib/dyno'
-import {buildCommand} from '../lib/helpers'
+import Dyno from '../lib/run/dyno'
+import {buildCommand} from '../lib/run/helpers'
 
 export default class RunConsole extends Command {
   static hidden = true
